Switched from PythonMagick to wand
Was having a ton of trouble with PythonMagick, there didn't seem to be any good documentation and it looked abandoned. Switched over to wand, but now I need to figure out how to do the galaxy maps. For now, they are simply black and white versions of the normal flag.