Added a maximum year for random flavor events
The random flavor events that are generated now can support a maximum year. If the start date is after the max year, they will not appear as a random event. This will cause events like the moon landing and invention of internet to not fire in late start dates where it should have happened already, such as in Millenium Dawn.